Fix #337: perform origin checks for workers in Fetch
Instead of having the constructor throw a SecurityError this will
result in a network error that will be reported as an error event on
the Worker/SharedWorker instance.

Although it seems like this change removes support for data URLs, that
in fact already happened when Fetch was integrated and we did not set
the same-origin data-URL flag. Which is in line with Chrome and the
stated goals of at least one Firefox engineer.